# services/auth-service/docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: LMS Auth Service API
  version: 1.0.0
  description: |
    Authentication and Authorization microservice for the LMS Platform.
    
    This service handles:
    - Organization and College registration and authentication
    - Student, Teacher, HOD, Dean, and Non-Teaching Staff authentication
    - Password reset and recovery flows
    - Session management with single-device login enforcement
    - Token generation and refresh using PASETO v4 with token rotation
    - Account lockout protection (brute-force prevention)
    - Password policy enforcement (12+ characters, complexity requirements)
    - Password history (prevents reusing last 5 passwords)
    - Rate limiting (global and per-endpoint)
    
    **Rate Limiting:**
    - Global rate limit: 100 requests per 15 minutes per IP
    - Login endpoints: 5 attempts per minute per IP
    - Rate limit exceeded returns HTTP 429
    
    **Account Lockout:**
    - After 5 failed login attempts, account is locked for 15 minutes
    - Locked accounts return HTTP 423
    - Lockout is tracked in Redis and database
    
  contact:
    name: LMS Platform Team
  license:
    name: ISC
servers:
  - url: http://localhost:8000/auth/api
    description: Kong Gateway (Recommended - Use this for all API calls)
  - url: http://localhost:4001/auth/api
    description: Direct Service (Internal/Development only)
...
paths:
...
  /api-keys:
    post:
      summary: Create a new API Key
      description: |
        Generate a new API key for system-to-system authentication.
        Only admins and college admins can create API keys.
      operationId: createApiKey
      tags:
        - API Keys
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Display name for the API key
                  example: "Mobile App Integration"
                description:
                  type: string
                  description: Optional description of the API key's purpose
                  example: "For mobile app to sync student data"
                scopes:
                  type: array
                  items:
                    type: string
                  description: Array of scopes/permissions (defaults to ['read:basic'])
                  example: ["read:basic", "read:students"]
      responses:
        '201':
          description: API Key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/ApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      summary: List API Keys
      description: Retrieve all API keys created by the authenticated user
      operationId: listApiKeys
      tags:
        - API Keys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  results:
                    type: integer
                    example: 3
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKeyListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api-keys/{id}:
    delete:
      summary: Revoke an API Key
      description: Revoke/disable an API key. The key can no longer be used for authentication.
      operationId: revokeApiKey
      tags:
        - API Keys
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The API Key ID to revoke
          schema:
            type: string
      responses:
        '204':
          description: API Key revoked successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
      description: |
        PASETO v4 public token authentication.
        
        Include the token in the Authorization header:
        ```
        Authorization: Bearer v4.public.eyJ...
        ```
        
        Tokens are automatically set as HTTP-only cookies on login,
        but you can also use the Authorization header for API calls.
  schemas:
...
    ApiKey:
      type: object
      description: Complete API Key response after creation
      properties:
        id:
          type: string
          description: Unique API Key identifier
          example: "key_xyz123abc"
        name:
          type: string
          description: Display name for the API key
          example: "Mobile App Integration"
        prefix:
          type: string
          description: Public prefix of the API key (safe to log)
          example: "key_xyz"
        hash:
          type: string
          description: Hashed API key (never shown after creation)
          example: "scrypt(...)$..."
        scopes:
          type: array
          items:
            type: string
          description: Array of permissions/scopes
          example: ["read:basic", "read:students"]
        isActive:
          type: boolean
          description: Whether the API key is currently active
          example: true
        createdBy:
          type: string
          description: User ID who created this API key
          example: "org_abc123"
        createdByType:
          type: string
          description: Role of the creator
          enum: ["admin", "college_admin", "organization"]
          example: "organization"
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of last use, null if never used
          example: "2026-02-26T10:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: When this API key was created
          example: "2026-02-20T15:45:00Z"
        revokedAt:
          type: string
          format: date-time
          nullable: true
          description: When this API key was revoked, null if still active
          example: null
    ApiKeyListItem:
      type: object
      description: API Key information for list responses
      properties:
        id:
          type: string
          description: Unique API Key identifier
          example: "key_xyz123abc"
        name:
          type: string
          description: Display name for the API key
          example: "Mobile App Integration"
        prefix:
          type: string
          description: Public prefix of the API key
          example: "key_xyz"
        scopes:
          type: array
          items:
            type: string
          description: Array of permissions/scopes
          example: ["read:basic"]
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of last use
          example: "2026-02-26T10:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: When this API key was created
          example: "2026-02-20T15:45:00Z"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error or invalid input
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  status:
                    type: string
                    example: "fail"
                  message:
                    type: string
                    example: "Validation failed"
                  totalErrors:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        message:
                          type: string
                        field:
                          type: string
                        expectedType:
                          type: string
          examples:
            validationError:
              value:
                status: "fail"
                message: "Validation failed"
                totalErrors: 2
                errors:
                  - path: "email"
                    message: "Invalid email address"
                    field: "email"
                    expectedType: "string"
                  - path: "password"
                    message: "Password must be at least 8 characters"
                    field: "password"
                    expectedType: "string"
    Unauthorized:
      description: Unauthorized - invalid credentials, missing token, or expired session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCredentials:
              value:
                success: false
                message: "Invalid email or password"
                error: "Invalid credentials"
            missingToken:
              value:
                success: false
                message: "Authentication required"
                error: "Missing authorization header"
            invalidToken:
              value:
                success: false
                message: "Token must be a valid PASETO v4 public token"
                error: "Invalid token format"
    Forbidden:
      description: Forbidden - user doesn't have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "You are not authorized to perform this action"
            error: "Forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            error: "Not Found"
...